Fix command button prompt lint failure
